fix: keep review artifact extraction focused
